mergefrom sipx 1-7455 Bring up to date relative to SIPfoundry Main.
But SipRedirectorCallerAlias hasn't been converted into a plugin yet.


git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/user_dworley_redirectors@7460 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXportLib/include/os/OsTimer.h
@@ -1,10 +1,16 @@
-//
-// Copyright (C) 2004, 2005, 2006 Pingtel Corp.
-//
-//
+// 
+// 
+// Copyright (C) 2005-2006 SIPez LLC.
+// Licensed to SIPfoundry under a Contributor Agreement.
+// 
+// Copyright (C) 2004-2006 SIPfoundry Inc.
+// Licensed by SIPfoundry under the LGPL license.
+// 
+// Copyright (C) 2004-2006 Pingtel Corp.
+// Licensed to SIPfoundry under a Contributor Agreement.
+// 
 // $$
-////////////////////////////////////////////////////////////////////////
-//////
+//////////////////////////////////////////////////////////////////////////////
 
 
 #ifndef _OsTimer_h_
@@ -328,7 +334,7 @@ class OsTimer : public UtlContainable
    /// Convert a Timer into an OsTime.
    inline static void cvtToOsTime(OsTime& out, Time in)
    {
-      OsTime temp(in / 1000000, in % 1000000);
+      OsTime temp((long)in / 1000000, (long)in % 1000000);
       out = temp;
    }
 
@@ -348,7 +354,7 @@ class OsTimer : public UtlContainable
    inline static int compareTimes(Time a, Time b)
    {
       // Interpret a and b as long long int's and take their difference.
-      return a - b;
+      return (int)(a - b);
    }
    ///< Returns > 0, 0, or < 0 according to whether a > b, a == b, or a < b.
